Refactor: break `match_fields` module in three.
The original match_fields now imports:
* match_fields_base
* match_fields_ipv6
