+++ v0x04/match_fields.py
@@ -7,64 +7,24 @@ eXtended Match Type-Length-Value (OXM TLV) element.
 This module provides high-level Python classes for OXM TLV fields in order to
 make the OF 1.3 match fields easy to use and to be coded.
 """
-from abc import ABC, abstractmethod
 
-from pyof.foundation.basic_types import HWAddress, IPAddress, IPv6Address
+from pyof.foundation.basic_types import HWAddress, IPAddress
 from pyof.v0x04.common.flow_match import OxmOfbMatchField, OxmTLV, VlanId
 
+# pylint: disable=unused-import
+from napps.kytos.of_core.v0x04.match_fields_base import (MatchField,
+                                                         MatchFieldFactory)
+from napps.kytos.of_core.v0x04.match_fields_ipv6 import (MatchEXTHDR,
+                                                         MatchICMPV6Code,
+                                                         MatchICMPV6Type,
+                                                         MatchIPv6Dst,
+                                                         MatchIPv6FLabel,
+                                                         MatchIPv6Src,
+                                                         MatchNDSLL,
+                                                         MatchNDTarget,
+                                                         MatchNDTLL)
+# pylint: enable=unused-import
 from napps.kytos.of_core.v0x04.utils import bytes_to_mask, mask_to_bytes
-
-
-class MatchField(ABC):
-    """Base class for match fields. Abstract OXM TLVs of python-openflow.
-
-    Just extend this class and you will be forced to define the required
-    low-level attributes and methods below:
-
-    * "name" attribute (field name to be displayed in JSON);
-    * "oxm_field" attribute (``OxmOfbMatchField`` enum);
-    * Method to return a pyof OxmTLV;
-    * Method to create an instance from an OxmTLV.
-    """
-
-    def __init__(self, value):
-        """Define match field value."""
-        self.value = value
-
-    @property
-    @classmethod
-    @abstractmethod
-    def name(cls):
-        """Define a name to be displayed in JSON.
-
-        It can be overriden just by a class attibute.
-        """
-
-    @property
-    @classmethod
-    @abstractmethod
-    def oxm_field(cls):
-        """Define this subclass ``OxmOfbMatchField`` value.
-
-        It can be overriden just by as a class attibute.
-        """
-
-    @abstractmethod
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-
-    @classmethod
-    @abstractmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-
-    def __eq__(self, other):
-        """Two objects are equal if their values are the same.
-
-        The oxm_field equality is checked indirectly when comparing whether
-        the objects are instances of the same class.
-        """
-        return isinstance(other, self.__class__) and other.value == self.value
 
 
 class MatchDLVLAN(MatchField):
@@ -670,184 +630,6 @@ class MatchARPTHA(MatchField):
         return cls(value)
 
 
-class MatchIPv6Src(MatchField):
-    """Match for IPv6 source."""
-
-    name = 'ipv6_src'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_SRC
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        ip_addr = IPv6Address(self.value)
-        value_bytes = ip_addr.pack()
-        if ip_addr.netmask < 128:
-            value_bytes += mask_to_bytes(ip_addr.netmask, 128)
-        return OxmTLV(oxm_field=self.oxm_field,
-                      oxm_hasmask=ip_addr.netmask < 128,
-                      oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        ip_address = IPv6Address()
-        ip_address.unpack(tlv.oxm_value)
-        addr_str = str(ip_address)
-        value = addr_str
-        if tlv.oxm_hasmask:
-            value = f'{addr_str}/{bytes_to_mask(tlv.oxm_value[16:], 128)}'
-        return cls(value)
-
-
-class MatchIPv6Dst(MatchField):
-    """Match for IPv6 destination."""
-
-    name = 'ipv6_dst'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_DST
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        ip_addr = IPv6Address(self.value)
-        value_bytes = ip_addr.pack()
-        if ip_addr.netmask < 128:
-            value_bytes += mask_to_bytes(ip_addr.netmask, 128)
-        return OxmTLV(oxm_field=self.oxm_field,
-                      oxm_hasmask=ip_addr.netmask < 128,
-                      oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        ip_address = IPv6Address()
-        ip_address.unpack(tlv.oxm_value)
-        addr_str = str(ip_address)
-        value = addr_str
-        if tlv.oxm_hasmask:
-            value = f'{addr_str}/{bytes_to_mask(tlv.oxm_value[16:], 128)}'
-        return cls(value)
-
-
-class MatchIPv6FLabel(MatchField):
-    """Match for IPv6 Flow Label."""
-
-    name = 'ipv6_flabel'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_FLABEL
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        try:
-            value = int(self.value)
-            mask = None
-            oxm_hasmask = False
-        except ValueError:
-            value, mask = map(int, self.value.split('/'))
-            oxm_hasmask = True
-        value_bytes = value.to_bytes(4, 'big')
-        if mask:
-            value_bytes += mask.to_bytes(4, 'big')
-        return OxmTLV(oxm_field=self.oxm_field,
-                      oxm_hasmask=oxm_hasmask,
-                      oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        value = int.from_bytes(tlv.oxm_value[:4], 'big')
-        if tlv.oxm_hasmask:
-            flabel_mask = int.from_bytes(tlv.oxm_value[4:], 'big')
-            value = f'{value}/{flabel_mask}'
-        return cls(value)
-
-
-class MatchICMPV6Type(MatchField):
-    """Match for ICMPV6 type."""
-
-    name = 'icmpv6_type'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_ICMPV6_TYPE
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        value_bytes = self.value.to_bytes(1, 'big')
-        return OxmTLV(oxm_field=self.oxm_field, oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        port = int.from_bytes(tlv.oxm_value, 'big')
-        return cls(port)
-
-
-class MatchICMPV6Code(MatchField):
-    """Match for ICMPV6 code."""
-
-    name = 'icmpv6_code'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_ICMPV6_CODE
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        value_bytes = self.value.to_bytes(1, 'big')
-        return OxmTLV(oxm_field=self.oxm_field, oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        priority = int.from_bytes(tlv.oxm_value, 'big')
-        return cls(priority)
-
-
-class MatchNDTarget(MatchField):
-    """Match for IPv6 ND Target."""
-
-    name = 'nd_tar'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_ND_TARGET
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        value_bytes = self.value.to_bytes(16, 'big')
-        return OxmTLV(oxm_field=self.oxm_field, oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        target = int.from_bytes(tlv.oxm_value, 'big')
-        return cls(target)
-
-
-class MatchNDSLL(MatchField):
-    """Match for IPv6 ND SLL."""
-
-    name = 'nd_sll'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_ND_SLL
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        value_bytes = self.value.to_bytes(6, 'big')
-        return OxmTLV(oxm_field=self.oxm_field, oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        sll = int.from_bytes(tlv.oxm_value, 'big')
-        return cls(sll)
-
-
-class MatchNDTLL(MatchField):
-    """Match for IPv6 ND TLL."""
-
-    name = 'nd_tll'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_ND_TLL
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        value_bytes = self.value.to_bytes(6, 'big')
-        return OxmTLV(oxm_field=self.oxm_field, oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        tll = int.from_bytes(tlv.oxm_value, 'big')
-        return cls(tll)
-
-
 class MatchMPLSLabel(MatchField):
     """Match for MPLS Label."""
 
@@ -934,38 +716,6 @@ class MatchPBBISID(MatchField):
         return cls(value)
 
 
-class MatchEXTHDR(MatchField):
-    """Match for IPv6 EXTHDR."""
-
-    name = 'v6_hdr'
-    oxm_field = OxmOfbMatchField.OFPXMT_OFB_IPV6_EXTHDR
-
-    def as_of_tlv(self):
-        """Return a pyof OXM TLV instance."""
-        try:
-            value = int(self.value)
-            mask = None
-            oxm_hasmask = False
-        except ValueError:
-            value, mask = map(int, self.value.split('/'))
-            oxm_hasmask = True
-        value_bytes = value.to_bytes(2, 'big')
-        if mask:
-            value_bytes += mask.to_bytes(2, 'big')
-        return OxmTLV(oxm_field=self.oxm_field,
-                      oxm_hasmask=oxm_hasmask,
-                      oxm_value=value_bytes)
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        value = int.from_bytes(tlv.oxm_value[:2], 'big')
-        if tlv.oxm_hasmask:
-            exhead_mask = int.from_bytes(tlv.oxm_value[2:], 'big')
-            value = f'{value}/{exhead_mask}'
-        return cls(value)
-
-
 class MatchMetadata(MatchField):
     """Match for table metadata."""
 
@@ -1028,43 +778,3 @@ class MatchTUNNELID(MatchField):
             tunnel_mask = int.from_bytes(tlv.oxm_value[8:], 'big')
             value = f'{value}/{tunnel_mask}'
         return cls(value)
-
-
-class MatchFieldFactory(ABC):
-    """Create the correct MatchField subclass instance.
-
-    As OF 1.3 has many match fields and there are many ways to (un)pack their
-    OxmTLV.oxm_value, this class does all the work of finding the correct
-    MatchField class and instantiating the corresponding object.
-    """
-
-    __classes = {}
-
-    @classmethod
-    def from_name(cls, name, value):
-        """Return the proper object from name and value."""
-        field_class = cls._get_class(name)
-        if field_class:
-            return field_class(value)
-        return None
-
-    @classmethod
-    def from_of_tlv(cls, tlv):
-        """Return the proper object from a pyof OXM TLV."""
-        field_class = cls._get_class(tlv.oxm_field)
-        if field_class:
-            return field_class.from_of_tlv(tlv)
-        return None
-
-    @classmethod
-    def _get_class(cls, name_or_field):
-        """Return the proper object from field name or OxmTLV.oxm_field."""
-        if not cls.__classes:
-            cls._index_classes()
-        return cls.__classes.get(name_or_field)
-
-    @classmethod
-    def _index_classes(cls):
-        for subclass in MatchField.__subclasses__():
-            cls.__classes[subclass.name] = subclass
-            cls.__classes[subclass.oxm_field] = subclass
